+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### vNext
 * Added `<Mutation />` component [#1520](https://github.com/apollographql/react-apollo/pull/1520)
+* HoC `props` result-mapping function now receives prior return value as second argument.
 * Fix errorPolicy when 'all' not passing data and errors
 
 ### 2.1.0-beta.2